1613 - Refactor filtering rule column functions to be usable in any job
