+++ projects/_03_independent_cqc/_02_clean/fargate/utils/clean_ct_outliers/null_values_exceeding_repetition_limit.py
@@ -1,6 +1,6 @@
 import polars as pl
 
-from projects._03_independent_cqc._02_clean.fargate.utils.filtering_utils import (
+from polars_utils.filtering_utils import (
     update_filtering_rule,
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
